# tests/openapi.json
{
  "components": {
    "schemas": {
      "AssignRequest": {
        "additionalProperties": false,
        "properties": {
          "assignee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          }
        },
        "required": [
          "assignee_id"
        ],
        "title": "AssignRequest",
        "type": "object"
      },
      "Category": {
        "enum": [
          "network",
          "access",
          "hardware",
          "software",
          "other"
        ],
        "title": "Category",
        "type": "string"
      },
      "CommentRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "internal": {
            "default": false,
            "title": "Internal",
            "type": "boolean"
          }
        },
        "required": [
          "body"
        ],
        "title": "CommentRequest",
        "type": "object"
      },
      "DeactivateRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "DeactivateRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Level": {
        "enum": [
          "high",
          "medium",
          "low"
        ],
        "title": "Level",
        "type": "string"
      },
      "Status": {
        "enum": [
          "new",
          "open",
          "in_progress",
          "pending",
          "resolved",
          "closed"
        ],
        "title": "Status",
        "type": "string"
      },
      "TicketRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Category"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "maxLength": 5000,
            "minLength": 10,
            "title": "Description",
            "type": "string"
          },
          "impact": {
            "$ref": "#/components/schemas/Level",
            "default": "medium"
          },
          "title": {
            "maxLength": 120,
            "minLength": 5,
            "title": "Title",
            "type": "string"
          },
          "urgency": {
            "$ref": "#/components/schemas/Level",
            "default": "medium"
          }
        },
        "required": [
          "title",
          "description"
        ],
        "title": "TicketRequest",
        "type": "object"
      },
      "TransitionRequest": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "to_status": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "required": [
          "to_status"
        ],
        "title": "TransitionRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Help Desk",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/tickets": {
      "get": {
        "operationId": "api_list_api_tickets_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Status"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "include_closed",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Closed",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "queue",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Queue"
            }
          },
          {
            "in": "query",
            "name": "assignee_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignee Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api List",
        "tags": [
          "api"
        ]
      },
      "post": {
        "operationId": "api_create_api_tickets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Create",
        "tags": [
          "api"
        ]
      }
    },
    "/api/tickets/{ticket_id}": {
      "get": {
        "operationId": "api_detail_api_tickets__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Detail",
        "tags": [
          "api"
        ]
      }
    },
    "/api/tickets/{ticket_id}/assign": {
      "post": {
        "operationId": "api_assign_api_tickets__ticket_id__assign_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Assign",
        "tags": [
          "api"
        ]
      }
    },
    "/api/tickets/{ticket_id}/comments": {
      "post": {
        "operationId": "api_comment_api_tickets__ticket_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Comment",
        "tags": [
          "api"
        ]
      }
    },
    "/api/tickets/{ticket_id}/transition": {
      "post": {
        "operationId": "api_transition_api_tickets__ticket_id__transition_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Transition",
        "tags": [
          "api"
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "api_users_api_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Users",
        "tags": [
          "api"
        ]
      }
    },
    "/api/users/{user_id}/deactivate": {
      "post": {
        "operationId": "api_deactivate_api_users__user_id__deactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeactivateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Deactivate",
        "tags": [
          "api"
        ]
      }
    },
    "/api/users/{user_id}/reactivate": {
      "post": {
        "operationId": "api_reactivate_api_users__user_id__reactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Reactivate",
        "tags": [
          "api"
        ]
      }
    }
  }
}